gyms/AlfworldGym/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "alfworldgym"
version = "1.0.0"
description = "A Gymnasium environment for alfworld household task completion"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "AlfworldGym Team", email = "team@alfworldgym.ai"},
]
keywords = ["reinforcement-learning", "gymnasium", "alfworld", "household-tasks", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Games/Entertainment :: Puzzle Games",
]
dependencies = [
    "gymnasium>=0.26.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "alfworld",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black>=22.0",
    "flake8>=4.0",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/yourusername/AlfworldGym"
Repository = "https://github.com/yourusername/AlfworldGym"
Documentation = "https://github.com/yourusername/AlfworldGym#readme"
"Bug Tracker" = "https://github.com/yourusername/AlfworldGym/issues"

[tool.setuptools.packages.find]
include = ["alfworldgym*"] 